:truck: [#3489] Merge openforms.locations into openforms.contrib.kadaster
It was only interfacing with the BAG API anyway, and now the geo
endpoints code is co-located. The existing endpoint still works, and
it was already deprecated earlier. In Open Forms 3.0 we can make the
breaking change.
