+++ src/openforms/contrib/kadaster/api/views.py
@@ -1,5 +1,7 @@
 import logging
+from functools import partial
 
+from django.core.cache import caches
 from django.utils.translation import gettext_lazy as _
 
 from drf_spectacular.types import OpenApiTypes
@@ -13,9 +15,11 @@ from openforms.api.serializers import ExceptionSerializer, ValidationErrorSerial
 from openforms.api.views.mixins import ListMixin
 from openforms.submissions.api.permissions import AnyActiveSubmissionPermission
 
-from ..clients import get_locatieserver_client
+from ..clients import get_bag_client, get_locatieserver_client
 from .serializers import (
     AddressSearchResultSerializer,
+    GetStreetNameAndCityViewInputSerializer,
+    GetStreetNameAndCityViewResultSerializer,
     LatLngSearchInputSerializer,
     LatLngSearchResultSerializer,
 )
@@ -23,11 +27,76 @@ from .serializers import (
 logger = logging.getLogger(__name__)
 
 
+ADDRESS_AUTOCOMPLETE_CACHE_TIMEOUT = (
+    60 * 60 * 24
+)  # 24 hours - address data does NOT update frequently
+
+
+def lookup_address(postcode: str, number: str):
+    with get_bag_client() as client:
+        return client.get_address(postcode, number)
+
+
+class AddressAutocompleteView(APIView):
+    """
+    Get the street name and city when given a postcode and house number.
+    """
+
+    authentication_classes = ()
+    permission_classes = [AnyActiveSubmissionPermission]
+
+    @extend_schema(
+        summary=_("Get a street name and city"),  # type: ignore
+        description=_(
+            "Get the street name and city for a given postal code and house number.\n\n"
+            "**NOTE** the `/api/v2/location/get-street-name-and-city/` endpoint will "
+            "be removed in v3. Use `/api/v2/geo/address-autocomplete/` instead."
+        ),  # type: ignore
+        responses=GetStreetNameAndCityViewResultSerializer,
+        parameters=[
+            OpenApiParameter(
+                "postcode",
+                OpenApiTypes.STR,
+                OpenApiParameter.QUERY,
+                description=_("Postal code of the address"),  # type: ignore
+                required=True,
+            ),
+            OpenApiParameter(
+                "house_number",
+                OpenApiTypes.NUMBER,
+                OpenApiParameter.QUERY,
+                description=_("House number of the address"),  # type: ignore
+                required=True,
+            ),
+        ],
+        deprecated=True,
+    )
+    def get(self, request, *args, **kwargs):
+        serializer = GetStreetNameAndCityViewInputSerializer(data=request.query_params)
+        serializer.is_valid(raise_exception=True)
+        data = serializer.validated_data
+        postcode, number = data["postcode"], data["house_number"]
+
+        # check the cache so we avoid hitting the remote API too often (and risk
+        # of being throttled, see #1832)
+        address_data = caches["default"].get_or_set(
+            key=f"BAG|get_address|{postcode}|{number}",
+            default=partial(lookup_address, postcode, number),
+            timeout=ADDRESS_AUTOCOMPLETE_CACHE_TIMEOUT,
+        )
+
+        if not address_data:
+            # If address is not found just return an empty response
+            return Response({})
+
+        return Response(GetStreetNameAndCityViewResultSerializer(address_data).data)
+
+
 @extend_schema(
-    summary=_("Get an adress based on coordinates"),
+    summary=_("Get an adress based on coordinates"),  # type: ignore
     description=_(
         "Get the closest address name based on the given longitude and latitude."
-    ),
+    ),  # type: ignore
     parameters=[
         OpenApiParameter(
             "lat",
@@ -77,13 +146,13 @@ class LatLngSearchView(APIView):
 
 
 @extend_schema(
-    summary=_("List address suggestions with coordinates."),
+    summary=_("List address suggestions with coordinates."),  # type: ignore
     description=_(
         "Get a list of addresses, ordered by relevance/match score of the input "
         "query. Note that only results having latitude/longitude data are returned.\n\n"
         "The results are retrieved from the configured geo search service, defaulting "
         "to the Kadaster location server."
-    ),
+    ),  # type: ignore
     parameters=[
         OpenApiParameter(
             "q",
