--- src/openforms/locations/api/urls.py
@@ -1,11 +1,0 @@
-from django.urls import path
-
-from .views import GetStreetNameAndCityView
-
-urlpatterns = [
-    path(
-        "get-street-name-and-city",
-        GetStreetNameAndCityView.as_view(),
-        name="get-street-name-and-city-list",
-    ),
-]
